mysite/requirements.txt
python-3.5.1
psycopg2-2.6.2
Django-1.10.3
